Address fourth round of review feedback on #138: extract ChoicesMixin
Checked case statements and case-generate statements as requested - both had the exact same
duplicated 'store choices list + set Parent' pattern already, each marked with pre-existing
'# TODO: move to parent or grandparent' / '# TODO: what about choices?' comments.

New ChoicesMixin (Base.py, metaclass=ExtendedType, mixin=True), applied to:
- Common.py: SelectedWaveform, SelectedExpression
- Sequential.py: SequentialCase (used by Case and OthersCase)
- Concurrent.py: ConcurrentCase (used by GenerateCase and OthersGenerateCase)

Found and fixed a real, pre-existing bug while doing this: SequentialCase.__init__ never actually
initialized self._choices at all (just a bare type annotation, '# TODO: what about choices?') -
OthersCase (which doesn't override __init__) would raise AttributeError the moment .Choices was
ever accessed. Same bug existed for OthersGenerateCase via ConcurrentCase. ChoicesMixin.__init__
always initializes the list (even when choices=None), fixing this for both classes.

Had to be careful with parameter ordering: pyGHDL.dom's OthersCase/OthersGenerateCase DOM wrappers
call super().__init__(statements) / super().__init__(declaredItems, statements, alternativeLabel)
positionally (no choices argument, since they don't have any to pass). Placed the new choices
parameter after the existing ones in SequentialCase/ConcurrentCase's signatures specifically to
avoid silently breaking these pre-existing positional call sites - verified this holds by running
pyGHDL.dom's full test suite (not just pyVHDLModel's) against this change, and by directly
constructing an others-case/others-generate-case and confirming .Choices returns [] instead of
crashing.

Full suite: 117 passed, no regressions. Cross-checked against pyGHDL.dom (merged, includes regular
case-statement/case-generate translation, not just the new assignment work): 69 passed, 5 skipped.
