+++ pyVHDLModel/Concurrent.py
@@ -40,7 +40,7 @@ from pyTooling.Decorators    import export, readonly
 from pyTooling.MetaClasses   import ExtendedType
 
 from pyVHDLModel.Base        import ModelEntity, LabeledEntityMixin, DocumentedEntityMixin, Range, BaseChoice, BaseCase, IfBranchMixin
-from pyVHDLModel.Base        import ElsifBranchMixin, ElseBranchMixin, AssertStatementMixin, BlockStatementMixin, WaveformElement
+from pyVHDLModel.Base        import ElsifBranchMixin, ElseBranchMixin, AssertStatementMixin, BlockStatementMixin, WaveformElement, ChoicesMixin
 from pyVHDLModel.Regions     import ConcurrentDeclarationRegionMixin
 from pyVHDLModel.Namespace   import Namespace
 from pyVHDLModel.Name        import Name
@@ -697,7 +697,7 @@ class RangedGenerateChoice(ConcurrentChoice):
 
 
 @export
-class ConcurrentCase(BaseCase, LabeledEntityMixin, ConcurrentDeclarationRegionMixin, ConcurrentStatementsMixin, AllowBlackboxMixin):
+class ConcurrentCase(BaseCase, LabeledEntityMixin, ConcurrentDeclarationRegionMixin, ConcurrentStatementsMixin, AllowBlackboxMixin, ChoicesMixin):
 	_namespace: Namespace
 
 	def __init__(
@@ -705,6 +705,7 @@ class ConcurrentCase(BaseCase, LabeledEntityMixin, ConcurrentDeclarationRegionMi
 		declaredItems:    Nullable[Iterable] = None,
 		statements:       Nullable[Iterable[ConcurrentStatement]] = None,
 		alternativeLabel: Nullable[str] = None,
+		choices:          Nullable[Iterable[BaseChoice]] = None,
 		allowBlackbox:    Nullable[bool] = None,
 		parent:           Nullable[ModelEntity] = None
 	) -> None:
@@ -721,12 +722,11 @@ class ConcurrentCase(BaseCase, LabeledEntityMixin, ConcurrentDeclarationRegionMi
 		ConcurrentDeclarationRegionMixin.__init__(self, declaredItems)
 		ConcurrentStatementsMixin.__init__(self, statements)
 		AllowBlackboxMixin.__init__(self, allowBlackbox)
+		ChoicesMixin.__init__(self, choices)
 
 
 @export
 class GenerateCase(ConcurrentCase):
-	_choices: List[ConcurrentChoice]
-
 	def __init__(
 		self,
 		choices:          Iterable[ConcurrentChoice],
@@ -736,19 +736,7 @@ class GenerateCase(ConcurrentCase):
 		allowBlackbox:    Nullable[bool] = None,
 		parent:           Nullable[ModelEntity] = None
 	) -> None:
-		super().__init__(declaredItems, statements, alternativeLabel, allowBlackbox, parent)
-
-		# TODO: move to parent or grandparent
-		self._choices = []
-		if choices is not None:
-			for choice in choices:
-				self._choices.append(choice)
-				choice.Parent = self
-
-	# TODO: move to parent or grandparent
-	@property
-	def Choices(self) -> List[ConcurrentChoice]:
-		return self._choices
+		super().__init__(declaredItems, statements, alternativeLabel, choices, allowBlackbox, parent)
 
 	def __str__(self) -> str:
 		return "when {choices} =>".format(choices=" | ".join(str(c) for c in self._choices))
